libudo: use predefined int types

+++ include/file-ops.h
@@ -28,12 +28,12 @@ struct udo_file_ops;
  */
 struct udo_file_ops_create_info
 {
-	const char        *fname;
-	unsigned long int size;
-	off_t             offset;
-	unsigned char     create_pipe : 1;
-	unsigned char     create_dir : 1;
-	unsigned char     protect : 1;
+	const char    *fname;
+	size_t        size;
+	off_t         offset;
+	unsigned char create_pipe : 1;
+	unsigned char create_dir : 1;
+	unsigned char protect : 1;
 };
